lexicon and subtitles working
